refactor: Reorganized files in Header folder for better structure
Improve code maintainability and readability.
