+++ app/layout/header/index.tsx
@@ -5,7 +5,7 @@ import ChevronDownIcon from '../../../public/assets/chevron-down.svg'
 import Link from 'next/link'
 import AnimeNavListHover from './components/AnimeNavListHover'
 import MangaNavListHover from './components/MangaNavListHover'
-import UserSideMenu from './components/UserSideMenu'
+import UserSideMenu from './components/User/UserSideMenu'
 import NavListMenu from './components/NavListMenu'
 import SearchContainer from './components/SearchContainer'
 import NewsNavListHover from './components/NewsNavListHover'
